requirements.txt
rich
openai
anthropic
google-generativeai
aiosqlite
chromadb
networkx
sqlalchemy
psutil
tree-sitter
pydantic
python-dotenv
pyyaml
